requirements.txt
discord.py>=2.3.2
openai>=1.12.0
chromadb>=0.4.22
python-dotenv>=1.0.0
langchain>=0.1.10
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-groq>=0.1.0
sentence-transformers>=2.2.2
tiktoken>=0.5.2

